rename copy-in-jsx -> text-in-jsx
