+++ .eslintplugin.js
@@ -1,5 +1,5 @@
 exports.rules = {
-  'copy-in-jsx': require('./scripts/eslint-plugin/copy_in_jsx'),
+  'text-in-jsx': require('./scripts/eslint-plugin/text_in_jsx'),
   i18n: require('./scripts/eslint-plugin/i18n'),
   'href-with-rel': require('./scripts/eslint-plugin/rel'),
   'require-license-header': require('./scripts/eslint-plugin/require_license_header'),
