+++ scripts/eslint-plugin/text_in_jsx.js
