--- src/example.tsx
@@ -1,21 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-import React, { FunctionComponent } from 'react';
-
-interface ExampleProps {
-  name: string;
-  age: number;
-}
-export const Example: FunctionComponent<ExampleProps> = ({ name, age }) => {
-  return (
-    <div>
-      My name is {name} and I am {age} years old.
-    </div>
-  );
-};
